Move ErrorOnNegativeTestCalculator to its own file
